+++ src/components/AddressEdit/AddressEdit.tsx
@@ -74,6 +74,7 @@ const AddressEdit: React.FC<AddressEditProps> = props => {
     "streetAddress1",
     "streetAddress2"
   ];
+
   const formErrors = getFormErrors<
     keyof AddressTypeInput,
     AccountErrorFragment | OrderErrorFragment