+++ src/fragments/types/OrderErrorFragment.ts
@@ -3,7 +3,7 @@
 // @generated
 // This file was automatically generated and should not be edited.
 
-import { OrderErrorCode } from "./../../types/globalTypes";
+import { OrderErrorCode, AddressTypeEnum } from "./../../types/globalTypes";
 
 // ====================================================
 // GraphQL fragment: OrderErrorFragment
@@ -13,4 +13,5 @@ export interface OrderErrorFragment {
   __typename: "OrderError";
   code: OrderErrorCode;
   field: string | null;
+  addressType: AddressTypeEnum | null;
 }